backend/requirements.txt
fastapi[standard]==0.135.1
sqlalchemy==2.0.48
psycopg2==2.9.11
alembic==1.18.4
celery[redis]==5.6.2
pydantic-settings==2.13.1
sentence-transformers==5.3.0
PyJWT==2.12.1
pgvector==0.4.2
rapidfuzz==3.14.3
lightgbm==4.6.0
scikit-learn==1.8.0
slowapi==0.1.9
bcrypt>=4.2,<6
